Code Refactoring + Added new entities for tracking score and different types of food.
