+++ src/main/java/org/saurabh/snake/board/SimpleBoard.java
@@ -1,6 +1,7 @@
-package org.saurabh.snake.entity;
+package org.saurabh.snake.board;
 
 import org.saurabh.snake.Coordinate;
+import org.saurabh.snake.entity.Direction;
 
 public class SimpleBoard implements Board{
     private int row;
