+++ src/main/java/org/saurabh/snake/food/Food.java
@@ -1,0 +1,8 @@
+package org.saurabh.snake.food;
+
+public interface Food {
+    public int getFoodPoints();
+    public String getFoodSymbol();
+
+    public int getIncrementalLength();
+}
