+++ src/main/java/org/saurabh/snake/food/NormalFood.java
@@ -1,0 +1,18 @@
+package org.saurabh.snake.food;
+
+public class NormalFood implements Food{
+    @Override
+    public int getFoodPoints() {
+        return 1;
+    }
+
+    @Override
+    public String getFoodSymbol() {
+        return ".";
+    }
+
+    @Override
+    public int getIncrementalLength() {
+        return 1;
+    }
+}
